a2a-vertex-agent/requirements.txt
google-adk>=1.0.0
a2a-sdk>=0.3.4
google-cloud-aiplatform[agent_engines,adk]>=1.112.0
python-dotenv>=1.0.0
requests>=2.31.0
starlette>=0.27.0
httpx>=0.27.0
